+++ lib/bloc_ease.dart
@@ -5,3 +5,4 @@ export 'four_state.consumer.dart';
 export 'four_state.listener.dart';
 export 'four_states.dart';
 export 'state_widgets.provider.dart';
+export 'bloc_ease_combined_state_listener.dart';